+++ src/main/java/seedu/address/logic/commands/RedoCommand.java
@@ -5,7 +5,7 @@ import static java.util.Objects.requireNonNull;
 import seedu.address.commons.core.Messages;
 import seedu.address.logic.commands.exceptions.CommandException;
 import seedu.address.model.Model;
-import seedu.address.model.StateNotFoundException;
+import seedu.address.model.version.StateNotFoundException;
 
 /**
  * Returns to a version of the application before the last execution of an UndoCommand.
