+++ src/main/java/seedu/address/model/Model.java
@@ -8,6 +8,7 @@ import seedu.address.commons.core.GuiSettings;
 import seedu.address.model.good.Good;
 import seedu.address.model.supplier.Supplier;
 import seedu.address.model.transaction.Transaction;
+import seedu.address.model.version.Versionable;
 
 /**
  * The API of the Model component.
