+++ src/main/java/seedu/address/model/TransactionHistory.java
@@ -7,6 +7,7 @@ import java.util.List;
 import javafx.collections.ObservableList;
 import seedu.address.model.transaction.Transaction;
 import seedu.address.model.transaction.UniqueTransactionList;
+import seedu.address.model.util.Copyable;
 
 /**
  * Wraps all data at the address-book level
