+++ src/main/java/seedu/address/model/util/Copyable.java
@@ -1,4 +1,4 @@
-package seedu.address.model;
+package seedu.address.model.util;
 
 /**
  * Represents an object that can produce its independent copy.
